When we have trouble with the time of diff runs the SUPPRESSBIGDIFFS env var lets us more thoroughly skip the problematic test cases. There are not many, and the problem only happens with certain VMs. 	modified:   DWARFTEST.sh
latest data
	modified:   RegressionTestsResults.ods

A useful script to examine the ALLdd file from a test run
if it takes too long.
    scripts/timereport.py  |grep notable
will focus your attention on things taking longer than
one might think. If all is well it will report a small
handfull of executions.  But when it reports 'shell'
issues it means executing shell commands is the issue,
and that means diff is taking a long time.
	modified:   scripts/timereport.py