This is a cleanup not usually used.
If running regressiontests in a clean directory
then this is unnecessary (there is nothing to
save anyway), but usable there.

If you run regressiontests in the regressiontests source
(Do Not Do That) this might be useful, but you will have to
delete the few lines at the beginning that refuse to let this
run in the regressiontests source directory.
	modified:   CLEANUP.sh